+++ dask_geomodeling/raster/sources.py
@@ -9,8 +9,8 @@ from osgeo import gdal, gdal_array
 
 from datetime import datetime, timedelta
 
+from dask import config
 from dask_geomodeling import utils
-from dask_geomodeling.settings import settings
 
 from .base import RasterBlock
 
@@ -289,7 +289,7 @@ class RasterFileSource(RasterBlock):
     """
 
     def __init__(self, url, time_first, time_delta):
-        url = utils.safe_file_url(url, settings["FILE_ROOT"])
+        url = utils.safe_file_url(url, config.get("geomodeling.root"))
         if isinstance(time_first, datetime):
             time_first = utils.dt_to_ms(time_first)
         else:
@@ -317,7 +317,7 @@ class RasterFileSource(RasterBlock):
         try:
             return self._gdal_dataset
         except AttributeError:
-            path = utils.safe_abspath(self.url, settings["FILE_ROOT"])
+            path = utils.safe_abspath(self.url, config.get("geomodeling.root"))
             self._gdal_dataset = gdal.Open(path)
             return self._gdal_dataset
 
@@ -447,7 +447,7 @@ class RasterFileSource(RasterBlock):
 
         # open the dataset
         url = process_kwargs["url"]
-        path = utils.safe_abspath(url, settings["FILE_ROOT"])
+        path = utils.safe_abspath(url, config.get("geomodeling.root"))
         dataset = gdal.Open(path)
         first_band = process_kwargs["first_band"]
         last_band = process_kwargs["last_band"]